\name{LuceVolume_indx}
\Rdversion{1.1}
\alias{LuceVolume_indx}
\docType{data}
\title{Index of Neighbors for kNN Classification of Volume, k = 18}
\description{
  An index of each time a data element was used as a nearest neighbor in
  a kNN classification where k = 18.
}

\usage{data(LuceVolume_indx)}

\format{
  A data frame with 671 rows and 20 columns.
}

\details{
  Column 1 represents the first neighbor chosen in the classification,
  column 2 the second chosen up to column 18 being the 18th neighbor chosen.
}

\keyword{datasets}
