\name{VOLUMEnnIndex_k20}
\Rdversion{1.1}
\alias{VOLUMEnnIndex_k20}
\docType{data}
\title{Index of neighbors for kNN classification of volume, k = 20}
\description{
  An index of each time a data element was used as a nearest neighbor in
  a kNN classification where k = 20.
}

\usage{data(VOLUMEnnIndex_k20)}

\format{
  A data frame with 671 rows and 20 columns.
}

\details{
  Column 1 represents the first neighbor chosen in the classification,
  column 2 the second chosen up to column 20 being the 20th neighbor chosen.
}

\keyword{datasets}
