\name{nnDiag-package}
\Rdversion{1.1}
\alias{nnDiag-package}
\alias{nnDiag}
\docType{package}

\title{
  A suite of graphical diagnostic tools to evaluate kNN classifications.
}

\description{
  Contains diagnostic functions that can be applied to k-nearest
  neighbor classifications, both continuous and categorical data.
}

\details{
  \tabular{ll}{
    Package: \tab nnDiag\cr
    Type: \tab Package\cr
    Version: \tab 0.0-5\cr
    Date: \tab 2010-01-11\cr
    License: \tab GPL (>= 2)\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Brian Walters \email{walte137@msu.edu}, Andrew O. Finley \email{finleya@msu.edu}, Zhen Zhang \email{zhangz19@stt.msu.edu}

  Maintainer: Brian Walters
}

\references{
  Hudson, W.D. and Ramm, C.W. (1987) Correct Formulation of the Kappa
  Coefficient of Agreement, \emph{Photogrammetric Engineering and Remote
    Sensing}. \bold{53}, 421--422.  
  
  McRoberts, R.E., Tomppo, E.O., Finley, A.O., Heikkinen, J. (2007)
  Estimating areal means and variances of forest attributes using the
  k-Nearest Neighbors technique and satellite imagery, \emph{Remote
    Sensing of Environment}. \bold{111}, 466--480.

  McRoberts, R.E. (2009) Diagnostic tools for nearest neighbors
  techniques when used with satellite imagery, \emph{Remote Sensing of
    Environment}. \bold{113}, 489--499.

  Rosenfield, G.H. and Fitzpatrick-Lins, K. (1986) A Coefficient of
  Agreement as a Measure of Thematic Classification Accuracy,
  \emph{Photogrammetric Engineering and Remote Sensing}. \bold{52},
  223--227.

  Story, M. and Congalton, R.G. (1986) Accuracy Assessment: A User's
  Perspective, \emph{Photogrammetric Engineering and Remote
  Sensing}. \bold{52}, 397--399.  
}

\keyword{package}
